sphinx-agent: Introduce `Esbonio` object
The `Esbonio` object, accessible via the `esbonio` field on the Sphinx
application object provides a single entry point from which we can
expose all our integrations with the parent language server.

This should also make it easy for extensions to detect when they are
running within `esbonio` and enable any additional integrations.

As well as exposing the database, this object now also holds our
logging handler.
